% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADASYN.R
\name{ADASYN}
\alias{ADASYN}
\title{Generate samples by ADASYN approach.}
\usage{
ADASYN(p, n, nTarget, k, m)
}
\arguments{
\item{p}{minority class samples}

\item{n}{majority class samples}

\item{nTarget}{the targeted number of samples to achieve}

\item{k}{is the number of nearest neighbours in the ADASYN algorithm, with the default value of 5}

\item{m}{seeds from the positive class in k-NN of the ADASYN algorithm, with the default value of 15}
}
\value{
sampleADA
}
\description{
Generate samples by ADASYN approach.
}
\keyword{internal}
