% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADASYNBar.R
\name{ADASYNBar}
\alias{ADASYNBar}
\title{Generate samples by ADASYN approach.}
\usage{
ADASYNBar(p, n, nTarget, k, m)
}
\arguments{
\item{p}{minority class samples}

\item{n}{majority class samples}

\item{nTarget}{the targeted number of samples to achieve}

\item{k}{number of nearest neighbours in k-NN search used by the ADASYN algorithm, with the default value of 5}

\item{m}{seeds from the positive class in m-NN search of the ADASYN algorithm, with the default value of 15}
}
\value{
sampleADA
}
\description{
Generate samples by ADASYN approach.
}
\keyword{internal}
