% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADASYNPara.R
\name{ADASYNPara}
\alias{ADASYNPara}
\title{Generate samples by ADASYN approach.}
\usage{
ADASYNPara(p, n, nTarget, k, m)
}
\arguments{
\item{p}{minority class samples}

\item{n}{majority class samples}

\item{nTarget}{the targeted number of samples to achieve}

\item{k}{number of the neareat neighbour in k-NN used by the ADASYN algorithm, with the default value of 5}

\item{m}{seeds from the positive class in m-NN used by the ADASYN algorithm, with the default value of 15}
}
\value{
sampleADA
}
\description{
Generate samples by ADASYN approach.
}
\keyword{internal}
