% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADASYNParaBar.R
\name{ADASYNParaBar}
\alias{ADASYNParaBar}
\title{Generate samples by ADASYN approach.}
\usage{
ADASYNParaBar(P, N, nTarget, k, m)
}
\arguments{
\item{P}{minority class samples}

\item{N}{majority class samples}

\item{nTarget}{the targeted number of samples to achieve}

\item{k}{k-NN used in the ADASYN algorithm, with the default value 5}

\item{m}{m-NN used in ADASYN, finding seeds from the Positive Class, with the default value 15}
}
\value{
sample_ada
}
\description{
Generate samples by ADASYN approach.
}
\keyword{internal}

