% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AfterLabelHFT.R
\name{AfterLabelHFT}
\alias{AfterLabelHFT}
\title{The precomputed prediction labels from LSTM for HFT after oversampling.}
\format{A list of prediction labels.}
\usage{
OSTOC::AfterLabelHFT()
}
\value{
AfterLabelHFT: the precomputed prediction labels from LSTM for HFT after oversampling
}
\details{
The LSTM for large dataset is slow. To avoid running LSTM every time creating 
the vignettes, the precomputed result is recorded for vignettes knit efficiency.
}
