% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AfterLabelMHEALTH.R
\name{AfterLabelMHEALTH}
\alias{AfterLabelMHEALTH}
\title{The precomputed prediction labels from LSTM for MHEALTH after oversampling.}
\format{A list of prediction labels.}
\usage{
OSTOC::AfterLabelMHEALTH()
}
\value{
AfterLabelMHEALTH: the precomputed prediction labels from LSTM for MHEALTH after oversampling
}
\details{
The LSTM for large dataset is slow. To avoid running LSTM every time creating 
the vignettes, the precomputed result is recorded for vignettes knit efficiency.
}
