% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BefoLabelHFT.R
\name{BefoLabelHFT}
\alias{BefoLabelHFT}
\title{The precomputed prediction labels from LSTM for HFT before oversampling.}
\format{A list of prediction labels.}
\usage{
OSTOC::BefoLabelHFT()
}
\value{
BefoLabelHFT: the precomputed prediction labels from LSTM for HFT before oversampling
}
\details{
The LSTM for large dataset is slow. To avoid running LSTM every time creating 
the vignettes, the precomputed result is recorded for vignettes knit efficiency.
}
