% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BefoLabelMHEALTH.R
\name{BefoLabelMHEALTH}
\alias{BefoLabelMHEALTH}
\title{The precomputed prediction labels from LSTM for MHEALTH before oversampling.}
\format{A list of prediction labels.}
\usage{
OSTOC::BefoLabelMHEALTH()
}
\value{
BefoLabelMHEALTH: the precomputed prediction labels from LSTM for MHEALTH before oversampling
}
\details{
The LSTM for large dataset is slow. To avoid running LSTM every time creating 
the vignettes, the precomputed result is recorded for vignettes knit efficiency.
}
