% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BefoResuMHEALTH.R
\name{BefoResuMHEALTH}
\alias{BefoResuMHEALTH}
\title{The precomputed result of LSTM for MHEALTH before oversampling.}
\format{A list of LSTM result. 
\describe{
  params: list of 7
  metrics: list of 4
  ...
}}
\usage{
OSTOC::BefoResuMHEALTH()
}
\value{
BefoResuMHEALTH: the precomputed result of LSTM for MHEALTH before oversampling
}
\details{
The LSTM for large dataset is slow. To avoid running LSTM every time creating 
the vignettes, the precomputed result is recorded for vignettes knit efficiency.
}
