% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_Synthetic_Control.R
\docType{data}
\name{Dataset_Synthetic_Control}
\alias{Dataset_Synthetic_Control}
\title{The synthetically generated control charts.}
\format{A dataset with 600 observations and a sequence length of 60, with a single sequence per line. 
\describe{
  The y data is the response (1 or 0).
  
  The x data constructs time series sequences (numeric).
  
  Both training and testing datasets contain 300 sequence observations.
}}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Synthetic+Control+Chart+Time+Series}
}
\usage{
data(Dataset_Synthetic_Control)
}
\description{
The data represents control charts synthetically generated by the process described in (Alcock and Manolopoulos, 1999).
}
\details{
Class 1 represents 'normal' status,  
while class 0 represents either 'Cyclic', 'Increasing trend', 'Decreasing trend', 'Upward shift' or Downward shift.
}
\keyword{datasets}
