% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPSO.R
\name{EPSO}
\alias{EPSO}
\title{Generate samples by EPSO algorithm.}
\usage{
EPSO(Me, V, D, P, N, R, M, NumToGen)
}
\arguments{
\item{Me}{Mean vector of positive class}

\item{V}{Eigen axes matrix (Each axis is a column vector)}

\item{D}{Modified Eigen Spectrum Value}

\item{P}{The minority class samples}

\item{N}{The majority class samples}

\item{R}{An scalar ratio to tell in which level (towards the boundary) we shall push our syntactic data, 
with the default value 1}

\item{M}{Scalar tells the reliable portion of the eigen spectrum}

\item{NumToGen}{The number of samples to be generated}
}
\value{
sample_epso
}
\description{
Generate samples by EPSO algorithm.
}
\keyword{internal}
