% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRatio.R
\name{FindRatio}
\alias{FindRatio}
\title{Find the distribution of the positive data.}
\usage{
FindRatio(P, N, m)
}
\arguments{
\item{P}{minority class samples}

\item{N}{majority class samples}

\item{m}{m-NN used in ADASYN, finding seeds from the Positive Class, with the default value 15}
}
\value{
ratio
}
\description{
Find the distribution of the positive data.
}
\keyword{internal}
