% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INOS2013.R
\name{INOS2013}
\alias{INOS2013}
\title{Generate samples by EPSO and ADASYN.}
\usage{
INOS2013(P, N, nTarget, R, Per, k, m)
}
\arguments{
\item{P}{minority class samples}

\item{N}{majority class samples}

\item{nTarget}{The targeted number of samples to achieve}

\item{R}{An scalar ratio to tell in which level (towards the boundary) we shall push our syntactic data, 
with the default value 1}

\item{Per}{Percentage of the mixing between SPO and ADASYN, with the default value 0.8}

\item{k}{k-NN used in the ADASYN algorithm, with the default value 5}

\item{m}{m-NN used in ADASYN, finding seeds from the Positive Class, with the default value 15}
}
\value{
myData
}
\description{
Generate samples by EPSO and ADASYN.
}
\keyword{internal}
