% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSTM_HFT_full_after.R
\docType{data}
\name{LSTM_HFT_full_after}
\alias{LSTM_HFT_full_after}
\title{The precomputed result of LSTM for HFT after oversampling.}
\format{A list of LSTM result. 
\describe{
  params: list of 7
  metrics: list of 4
  ...
}}
\usage{
data(LSTM_HFT_full_after)
}
\details{
The LSTM for 30000 observations dataset is slow. To avoid running LSTM every time creating 
the vignettes, the precomputed result is recorded for vignettes knit efficiency.
}
\keyword{datasets}
