% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSTM_MHEALTH_full_before.R
\docType{data}
\name{LSTM_MHEALTH_full_before}
\alias{LSTM_MHEALTH_full_before}
\title{The precomputed result of LSTM for MHEALTH before oversampling.}
\format{A list of LSTM result. 
\describe{
  params: list of 7
  metrics: list of 4
  ...
}}
\usage{
data(LSTM_MHEALTH_full_before)
}
\details{
The LSTM for large dataset is slow. To avoid running LSTM every time creating 
the vignettes, the precomputed result is recorded for vignettes knit efficiency.
}
\keyword{datasets}
