% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSTMLabel.R
\name{LSTMLabel}
\alias{LSTMLabel}
\title{The precomputed prediction labels from LSTM.}
\usage{
LSTMLabel(labelData)
}
\arguments{
\item{labelData}{the precomputed prediction labels from LSTM}
}
\value{
resu: the precomputed prediction labels in vector format
}
\description{
The LSTM for large dataset is slow. To avoid running LSTM every time creating 
the vignettes, the precomputed result is recorded for vignettes knit efficiency.
}
