% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSTMResu.R
\name{LSTMResu}
\alias{LSTMResu}
\title{The precomputed result of LSTM.}
\usage{
LSTMResu(fullData)
}
\arguments{
\item{fullData}{the precomputed result of LSTM}
}
\value{
resu: the precomputed result of LSTM in class keras_training_history
}
\description{
The LSTM for large dataset is slow. To avoid running LSTM every time creating 
the vignettes, the precomputed result is recorded for vignettes knit efficiency.
}
