% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_control.R
\docType{data}
\name{synthetic_control}
\alias{synthetic_control}
\title{The synthetically generated control charts.}
\format{A data frame with 555 observations and 61 variables, with a single chart per line. 
\describe{
  The first column is the class label (1 or 0)
  The other 60 columns constructs a  time series sequence (in numeric)
  ...
}}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Synthetic+Control+Chart+Time+Series}
}
\usage{
data(synthetic_control)
}
\description{
The data is generated by the process in Alcock and Manolopoulos (1999). Class 1 aims to Normal status,  
while class 0 aims to Cyclic, Increasing trend, Decreasing trend, Upward shift and Downward shift.
}
\keyword{datasets}

