% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset_Synthetic_Control.R
\docType{data}
\name{Dataset_Synthetic_Control}
\alias{Dataset_Synthetic_Control}
\title{The synthetically generated control charts.}
\format{A dataset with 600 observations and 60 sequence length, with a single chart per line. 
\describe{
  The y data is the class label (1 or 0)
  The x data constructs time series sequences (in numeric)
  Both training and testing datasets contain 300 observations
  ...
}}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Synthetic+Control+Chart+Time+Series}
}
\usage{
data(Dataset_Synthetic_Control)
}
\details{
The data is generated by the process in Alcock and Manolopoulos (1999). Class 1 aims to Normal status,  
while class 0 aims to Cyclic, Increasing trend, Decreasing trend, Upward shift and Downward shift.
}
\keyword{datasets}
