% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESPO.R
\name{ESPO}
\alias{ESPO}
\title{Generate samples by ESPO algorithm.}
\usage{
ESPO(me, v, dMod, p, n, r, m, numToGen)
}
\arguments{
\item{me}{Mean vector of positive class}

\item{v}{Eigen axes matrix (Each axis is a column vector)}

\item{dMod}{Modified Eigen Spectrum Value}

\item{p}{The minority class samples}

\item{n}{The majority class samples}

\item{r}{An scalar ratio to tell in which level (towards the boundary) we shall push our syntactic data, 
with the default value 1}

\item{m}{Scalar tells the reliable portion of the eigen spectrum}

\item{numToGen}{The number of samples to be generated}
}
\value{
sampleESPO
}
\description{
Generate samples by ESPO algorithm.
}
\keyword{internal}

