% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReguCovar.R
\name{ReguCovar}
\alias{ReguCovar}
\title{Generate samples by ESPO and ADASYN.}
\usage{
ReguCovar(cleanData, target_class, ratio, R, Per, k, m, parallel, progBar)
}
\arguments{
\item{cleanData}{First column is label data, rest is sample data, without NA, NaN values}

\item{target_class}{The class needs to be oversampled}

\item{ratio}{Targeted positive samples number to achieve/negative samples number, with the default value 1.}

\item{R}{An scalar ratio to tell in which level (towards the boundary) we shall push our syntactic data, 
with the default value 1}

\item{Per}{Percentage of the mixing between ESPO and ADASYN, with the default value 0.8}

\item{k}{k-NN used in the ADASYN algorithm, with the default value 5}

\item{m}{m-NN used in ADASYN, finding seeds from the Positive Class, with the default value 15}

\item{parallel}{Whether to run in parallel, with the default setting TRUE 
(Recommend for dataset with over 30,000 records)}

\item{progBar}{Whether to include progress bars, with the default setting TRUE}
}
\value{
newData
}
\description{
Generate samples by ESPO and ADASYN.
}
\keyword{internal}

